+++ python/common/animate.py
