+++ python/common/symbolic_dynamics.py
