+++ python/formulator/dynamics_base.py
